requirements.txt
astroid==2.2.5
autopep8==1.4.3
Django==1.11
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
selenium==3.141.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
psycopg2>=2.7,<3.0
